Add LLVM pass for sanity checks (#470)
* Fix call-callee type mismatches and mismapped function

* Fix return type mismatch on __chip_rnorm_f64().

* Fix call to scalbln(double, long) in OCML but it does not define one.

* Fix call to scalb(double, int) in OCML but second argument type is
  incorrect.

These were catched by HipSanityChecks pass.

* Add LLVM pass for sanity checks

Starting with checks for

* indirect calls which are not supported yet.

* Caller and callee disagrees on the the function type. (E.g. there is
  a call to `foo(double)` but the linked-in definition is actually
  `foo(float)`).

+++ bitcode/devicelib.cl
@@ -346,7 +346,7 @@ EXPORT float __chip_norm_f32(int dim, const float *a) {
   return sqrt(r);
 }
 
-  EXPORT float __chip_rnorm_f64(int dim, const double *a) { 
+  EXPORT double __chip_rnorm_f64(int dim, const double *a) {
   double r = 0;
   while (dim--) {
     r += a[0] * a[0];
